Rename kas::shell -> kas::app
