+++ crates/kas-core/src/app/common.rs
