+++ crates/kas-core/src/app/event_loop.rs
