+++ examples/clock.rs
@@ -18,14 +18,14 @@ use std::f32::consts::PI;
 use std::str::FromStr;
 use std::time::Duration;
 
+use kas::app::AppAssoc;
 use kas::draw::color::{Rgba, Rgba8Srgb};
 use kas::draw::{Draw, DrawRounded};
 use kas::geom::{Offset, Quad, Rect, Vec2};
 use kas::prelude::*;
-use kas::shell::AppAssoc;
 use kas::text::Text;
 
-type Application = kas::shell::Default<(), kas::theme::FlatTheme>;
+type Application = kas::app::Default<(), kas::theme::FlatTheme>;
 
 impl_scope! {
     #[derive(Clone)]
@@ -171,7 +171,7 @@ impl_scope! {
     }
 }
 
-fn main() -> kas::shell::Result<()> {
+fn main() -> kas::app::Result<()> {
     env_logger::init();
 
     let window = Window::new(Clock::new(), "Clock")
