+++ examples/data-list-view.rs
@@ -208,7 +208,7 @@ impl Driver<Item, Data> for MyDriver {
     }
 }
 
-fn main() -> kas::shell::Result<()> {
+fn main() -> kas::app::Result<()> {
     env_logger::init();
 
     let controls = row![
@@ -243,5 +243,5 @@ fn main() -> kas::shell::Result<()> {
 
     let window = Window::new(ui, "Dynamic widget demo");
 
-    kas::shell::Default::new(())?.with(window).run()
+    kas::app::Default::new(())?.with(window).run()
 }
